refactor: pages 분리
