+++ src/advanced/__tests__/advanced.test.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import { describe, expect, test } from 'vitest';
 import { act, fireEvent, render, screen, within } from '@testing-library/react';
-import { CartPage } from '../../refactoring/components/CartPage';
+import { CartPage } from '../../refactoring/pages/CartPage';
 import { AdminPage } from '../../refactoring/components/admin/AdminPage';
 import { Coupon, Product } from '../../types';
