+++ src/basic/__tests__/basic.test.tsx
@@ -8,7 +8,7 @@ import {
   screen,
   within,
 } from '@testing-library/react';
-import { CartPage } from '../../refactoring/components/CartPage';
+import { CartPage } from '../../refactoring/pages/CartPage';
 import { AdminPage } from '../../refactoring/components/admin/AdminPage';
 import { CartItem, Coupon, Product } from '../../types';
 import { useCart, useCoupons, useProducts } from '../../refactoring/hooks';
